genPatientID, newPatient, comboBox functionality
genPatientID - generates an ID for patient using first initial of firstName and firstInitial of lastName + a random int between 1000 & 9999

newPatient - Creates a new PatientRecord in the DB based on input. Also creates a new row in UserType in DB and a new row in Login.

Methods to generate the list of doctors for the combo boxes in Nurse.FXML implemented

All components lightly tested.